+++ CHANGELOG.md
@@ -7,11 +7,14 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - `defaultTeeFramework` option for `IExecConfig` and `chain.json` for choosing the default TEE framework to use
+- `iexec app init` option `--tee-framework <'scone'|'gramine'>` to specify the TEE framework to use for the app
 
 ### Changed
 
 - SMS URL resolution depends on TEE framework (default `scone`)
 - SMS URL override in `IExecConfig` or `chain.json` accepts `Record<TeeFramework,Url> | string`
+- `iexec app push-secret` & `app.pushAppSecretExists(appAddress)` use a TEE framework inferred from app if not specified
+- `iexec app check-secret` & `app.checkAppSecret(appAddress, secret)` use a TEE framework inferred from app if not specified
 
 ## [7.2.2] 2022-09-05